Split out OpenAI functions
